+++ Realm/RLMUtil.mm
@@ -29,6 +29,7 @@
 #import "RLMSchema_Private.h"
 #import "RLMSwiftSupport.h"
 #import "RLMUUID_Private.hpp"
+#import "RLMValue.h"
 
 #import <realm/mixed.hpp>
 #import <realm/object-store/shared_realm.hpp>
@@ -45,64 +46,6 @@
 #if !defined(REALM_COCOA_VERSION)
 #import "RLMVersion.h"
 #endif
-
-static inline bool numberIsInteger(__unsafe_unretained NSNumber *const obj) {
-    char data_type = [obj objCType][0];
-    return data_type == *@encode(bool) ||
-           data_type == *@encode(char) ||
-           data_type == *@encode(short) ||
-           data_type == *@encode(int) ||
-           data_type == *@encode(long) ||
-           data_type == *@encode(long long) ||
-           data_type == *@encode(unsigned short) ||
-           data_type == *@encode(unsigned int) ||
-           data_type == *@encode(unsigned long) ||
-           data_type == *@encode(unsigned long long);
-}
-
-static inline bool numberIsBool(__unsafe_unretained NSNumber *const obj) {
-    // @encode(BOOL) is 'B' on iOS 64 and 'c'
-    // objcType is always 'c'. Therefore compare to "c".
-    if ([obj objCType][0] == 'c') {
-        return true;
-    }
-
-    if (numberIsInteger(obj)) {
-        int value = [obj intValue];
-        return value == 0 || value == 1;
-    }
-
-    return false;
-}
-
-static inline bool numberIsFloat(__unsafe_unretained NSNumber *const obj) {
-    char data_type = [obj objCType][0];
-    return data_type == *@encode(float) ||
-           data_type == *@encode(short) ||
-           data_type == *@encode(int) ||
-           data_type == *@encode(long) ||
-           data_type == *@encode(long long) ||
-           data_type == *@encode(unsigned short) ||
-           data_type == *@encode(unsigned int) ||
-           data_type == *@encode(unsigned long) ||
-           data_type == *@encode(unsigned long long) ||
-           // A double is like float if it fits within float bounds or is NaN.
-           (data_type == *@encode(double) && (ABS([obj doubleValue]) <= FLT_MAX || isnan([obj doubleValue])));
-}
-
-static inline bool numberIsDouble(__unsafe_unretained NSNumber *const obj) {
-    char data_type = [obj objCType][0];
-    return data_type == *@encode(double) ||
-           data_type == *@encode(float) ||
-           data_type == *@encode(short) ||
-           data_type == *@encode(int) ||
-           data_type == *@encode(long) ||
-           data_type == *@encode(long long) ||
-           data_type == *@encode(unsigned short) ||
-           data_type == *@encode(unsigned int) ||
-           data_type == *@encode(unsigned long) ||
-           data_type == *@encode(unsigned long long);
-}
 
 static inline RLMArray *asRLMArray(__unsafe_unretained id const value) {
     return RLMDynamicCast<RLMArray>(value) ?: RLMDynamicCast<RLMListBase>(value)._rlmArray;
@@ -188,7 +131,7 @@ BOOL RLMValidateValue(__unsafe_unretained id const value,
         case RLMPropertyTypeData:
             return [value isKindOfClass:[NSData class]];
         case RLMPropertyTypeAny:
-            return NO;
+            return [value conformsToProtocol:@protocol(RLMValue)];
         case RLMPropertyTypeLinkingObjects:
             return YES;
         case RLMPropertyTypeObject: {